Fix
Fixed a small bug relating to adding a depolarizing map after the first hadamard on the ancilla. Previously, the code didn't add this noise because the first hadamard is in a moment that is part of the initial state generation. We don't add noise to the initial state since this is the random input state. Thus, previously the first hadamard on the ancilla was noiseless. Depending on the operations after the third moment we could end up having double the noise on this first hadamard, but it's not a big problem. It is better to test the mitigation against more noise.