Refactoring
Completed searching for checks backwards. Now, we set p2 and push it through the circuit to get p1. By starting with p2, we are now able to feed the highest weights first and adjust the possible optimal solution p2 weight accordingly. The possible optimal p2 solution is bounded by the current max pauli weight of the set of paulis about to be processed, i.e., assigned to p2 and pushed through to get p1.