Feature
Added functions in circgenerator.py for getting a list of random circuits, a function that generates qasm files of random circuits, and a function that does both. These circuits have no initial state or checks. I pushed the functionality of gen_rand_circs.py into a function called generate_circuits_with_checks_to_files in circgenerator.py. This function creates the circuits and finds the checks, but it doesn't save the original circuit with no checks or initial state as a qasm file.