Fix
Originally, we didn't transpile the swap gate to cnot before counting the number of cnot gates. This results in miscounting the number of cnots for randomly generated circuits. We removed the swap gate from randomly generated circuits to fix this error. This fine to do because we are assuming fully connected qubits.